\name{lkrt-package}
\alias{lkrt-package}
\alias{lkrt}
\docType{package}
\title{
Automatically give labels to various response option formats.
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab lkrt\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-12-09\cr
License: \tab What license is it under?\cr
}

}
\author{
Seth Berry
Michael Clark
Maintainer: Who to complain to <yourfault@somewhere.net>

}

\keyword{ package }
